@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.5.5</version>
+        <version>3.5.6</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.swp.project</groupId>
@@ -27,8 +27,9 @@
         <url/>
     </scm>
     <properties>
-        <java.version>17</java.version>
-        <spring-ai.version>1.0.1</spring-ai.version>
+        <java.version>21</java.version>
+        <spring-ai.version>1.0.3</spring-ai.version>
+        <spring-grpc.version>0.11.0</spring-grpc.version>
     </properties>
     <dependencies>
 
@@ -99,6 +100,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
+            <groupId>org.springframework.grpc</groupId>
+            <artifactId>spring-grpc-spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.ai</groupId>
             <artifactId>spring-ai-starter-model-vertex-ai-embedding</artifactId>
         </dependency>
@@ -108,7 +113,12 @@
         </dependency>
         <dependency>
             <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-vector-store-pgvector</artifactId>
+            <artifactId>spring-ai-starter-vector-store-pinecone</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.google.protobuf</groupId>
+            <artifactId>protobuf-java</artifactId>
+            <version>4.32.1</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.ai</groupId>
@@ -126,6 +136,13 @@
                 <groupId>org.springframework.ai</groupId>
                 <artifactId>spring-ai-bom</artifactId>
                 <version>${spring-ai.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.springframework.grpc</groupId>
+                <artifactId>spring-grpc-dependencies</artifactId>
+                <version>${spring-grpc.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
