@@ -1,11 +1,8 @@
 package com.swp.project.initializer;
 
-
 import com.swp.project.service.CustomerAiService;
-import com.swp.project.service.product.ProductService;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
-
 
 import com.swp.project.service.user.AdminService;
 import com.swp.project.service.user.CustomerService;
@@ -24,9 +21,7 @@ public class DataInitializer implements CommandLineRunner {
     private final ManagerService managerService;
     private final SellerService sellerService;
     private final ShipperService shipperService;
-    private final ProductService productService;
     private final CustomerAiService customerAiService;
-
 
     @Override
     public void run(String... args) {
@@ -36,6 +31,6 @@ public class DataInitializer implements CommandLineRunner {
         shipperService.initShipper();
         customerService.initCustomer();
 
-//        productService.getAllProducts().forEach(customerAiService::saveProductToVectorStore);
+//        customerAiService.saveAllProductsFromDatabaseToVectorStore();
     }
 }
